refactor(dev): move the MCP tool schemas out of the Worker package into the rs_dev adapter
ADR 0148 puts the Worker BODY in `@reddb-io/worker` and the Plugin MCP in the
plugin: the schemas red-castle carried under `src/mcp/` were never Worker code,
they were the surface a host mounts. They move to `apps/dev/src/mcp-tools/`,
which is where the stateless ACP client that publishes them already lives, and
the package drops its `./mcp-server` export.

`apps/dev/src/mcp/` is deliberately NOT the destination: the retired-authority
guard reserves that path for the stateful Dev MCP implementation ADR 0144
deleted, and a `project.ts` landing back on the same path would make the ratchet
unreadable.

`extinct-nouns.ts` travels with them — `refuseFleetNaming` is a schema-level
refusal for an extinct input field, and only the tool schemas ever called it, so
the engine re-export it reached through had no other consumer. Its message names
the `rs_dev` `project_start` inputs now instead of a CLI flag on a binary ADR
0147 is deleting.

The server is named `rs_dev` (ADR 0147 rule 2), pinned beside the frozen tool
table: the name is part of the published surface, since a host addresses every
tool through it. The tool set is byte-identical.

Refs #4023
