+++ apps/dev/src/mcp-tools/claim.ts
