+++ apps/dev/src/mcp-tools/deadend.ts
