+++ apps/dev/src/mcp-tools/extinct-nouns.ts
@@ -2,7 +2,7 @@
 export const FLEET_REMOVED_MESSAGE =
   "named fleets were removed (ADR 0130): the host daemon owns the budget and each project has exactly one demand producer, " +
   "so a fleet name addresses nothing. Drop the name — the work scope, the runner and the base branch now ride with the " +
-  "project's producer: pass --selector / --runner / --base to `red-skills-dev fleet`, or call the redskilled `project_start` tool " +
+  "project's producer: pass `selector` / `runner` / `base` to the rs_dev `project_start` tool " +
   "(status: `project_status`, resize: `project_resize`, stop: `project_stop`).";
 
 /** Raised when an invocation still addresses a fleet by name. */
