+++ apps/dev/src/mcp-tools/help.ts
