+++ apps/dev/src/mcp-tools/hygiene.ts
