+++ apps/dev/src/mcp-tools/statusline.ts
