+++ apps/dev/src/mcp-tools/worker.ts
