+++ apps/dev/tests/mcp-extinct-nouns.test.ts
@@ -3,7 +3,7 @@ import {
   FLEET_REMOVED_MESSAGE,
   FleetNamingRemovedError,
   refuseFleetNaming,
-} from "./extinct-nouns.js";
+} from "../src/mcp-tools/extinct-nouns.js";
 
 describe("naming a fleet after its removal", () => {
   it("passes an absent name through, so a call site can guard unconditionally", () => {
@@ -21,12 +21,12 @@ describe("naming a fleet after its removal", () => {
       const message = (err as Error).message;
       expect(message).toContain('"nightly"');
       expect(message).toContain("project_start");
-      expect(message).toContain("--selector");
+      expect(message).toContain("`selector`");
     }
   });
 
   it("states the replacement route rather than only the removal", () => {
-    for (const route of ["project_status", "project_resize", "project_stop", "--runner", "--base"]) {
+    for (const route of ["project_status", "project_resize", "project_stop", "`runner`", "`base`"]) {
       expect(FLEET_REMOVED_MESSAGE).toContain(route);
     }
   });
