--- packages/worker/src/engine/extinct-nouns.ts
@@ -1,3 +1,0 @@
-// Compatibility path for engine consumers. The migration guard itself is
-// engine-free so static MCP metadata can import it without loading the engine.
-export * from "../extinct-nouns.js";
